tests/test_float.c: Pacify Coverity 4980336, overflow
Could not happen, but shut up Coverity...

+++ tests/test_float.c
@@ -29,7 +29,7 @@
 
 static int test_single(void) {
     static float f;
-    int i;
+    long long i;   // Pacify Coverity 4980336, overflow
     int e = 0;
 
     // addition test
